update dependencies, re-create example
On `main` branch (at SHA), `yarn install` raises several warnings about incorrect peer dependencies and the subsequently triggered build fails with:
> (rpt2 plugin) Error: /Users/cchudzicki/dev/react-mathlive/src/index.tsx(28,7): semantic error TS2322: Type '((event: FocusEvent<MathfieldElement>) => void) | undefined' is not assignable to type 'string | number | boolean | undefined'.

The build error was resolved by updating to the most recent version of `mathlive` (which required changing a few things, e.g., "MathfieldConfig" was renamed to "MathfieldOptions" somewhere between mathlive 0.60 and 0.73).

This commit also updates several other packages, notably declaring the peer dependencies as react ^17 | ^18. The previous peer dependency of ^16.0.0 was incorrect since hooks were introduced in React 16.8.

In the example app, React was throwing an error:
> plugin "react" was conflicted between
which I resolved by re-creating with create-react-app. This had the benefit of updating the example to react 18.

Previously the example app was using npm links for all of its dependencies. I switched this to only use npm link for math-live and its peer dependencies. This means that some dev-deps could be removed from the main package, which is nice. (Not too consequential now, but could be useful for reducing install time on CI.)